pyproject.toml
[project]
name = "blood-donor-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.3",
    "celery>=5.6.2",
    "fastapi[all]>=0.128.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]
